Adding logo tag to readme as requested in issue 5
What: a request was made to update the readme to include a logo
tag when running outside a node environment.
How: I've updated the readme to include the logo tag.
Testing: I tried to reproduce this while running under a nginx server and could not.
However, since the logo tag did not break anything I am including it in the readme.
Closes: issue 5